requirements.txt
fastapi
uvicorn
pandas
python-multipart
psycopg2-binary
